requirements.txt
django==1.11
djangorestframework==3.6.2